Refactor k8 types into packages
