+++ pkg/k8shandler/factory/daemonset.go
@@ -1,0 +1,46 @@
+package factory
+
+import (
+	apps "k8s.io/api/apps/v1"
+	core "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+//NewDaemonSet stubs an instance of a daemonset
+func NewDaemonSet(daemonsetName, namespace, loggingComponent, component string, podSpec core.PodSpec) *apps.DaemonSet {
+	labels := map[string]string{
+		"provider":      "openshift",
+		"component":     component,
+		"logging-infra": loggingComponent,
+	}
+	return &apps.DaemonSet{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "DaemonSet",
+			APIVersion: apps.SchemeGroupVersion.String(),
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      daemonsetName,
+			Namespace: namespace,
+			Labels:    labels,
+		},
+		Spec: apps.DaemonSetSpec{
+			Selector: &metav1.LabelSelector{
+				MatchLabels: labels,
+			},
+			Template: core.PodTemplateSpec{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:   daemonsetName,
+					Labels: labels,
+					Annotations: map[string]string{
+						"scheduler.alpha.kubernetes.io/critical-pod": "",
+					},
+				},
+				Spec: podSpec,
+			},
+			UpdateStrategy: apps.DaemonSetUpdateStrategy{
+				Type:          apps.RollingUpdateDaemonSetStrategyType,
+				RollingUpdate: &apps.RollingUpdateDaemonSet{},
+			},
+		},
+	}
+}
